Fix #71 and #78
the search for a user now makes use of the more generic get_user function and does not rely on forms. This is a nicer UX as it is enough to put in any information known about the user (e.g. a first name which is only included in the username). It also is a tested function that therefore solves #78